+++ src/conditionalLogicService/types.ts
@@ -1,7 +1,7 @@
 import { FormTypes, SubmissionTypes } from '@oneblink/types'
 
 export type FormElementsCtrl = {
-  model: SubmissionTypes.S3SubmissionData['submission']
+  model: SubmissionTypes.S3SubmissionData['submission'] | undefined
   flattenedElements: FormTypes.FormElement[]
   parentFormElementsCtrl?: FormElementsCtrl
 }